+++ XCDYouTubeKit Demo/iOS Demo/NowPlayingInfoCenterProvider.h
@@ -1,0 +1,9 @@
+//
+//  Copyright (c) 2013-2014 Cédric Luthi. All rights reserved.
+//
+
+@interface NowPlayingInfoCenterProvider : NSObject
+
+@property (nonatomic, assign, getter = isEnabled) BOOL enabled; // defaults to `YES`
+
+@end