+++ XCDYouTubeKit Tests/XCDYouTubeClientTestCase.m
@@ -13,19 +13,21 @@
 
 - (void) testThatVideoIsAvailalbeOnDetailPageEventLabel
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"dQw4w9WgXcQ" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"dQw4w9WgXcQ" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video);
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testThatVideoHasMetadata
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"9TTioMbNT9I" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"9TTioMbNT9I" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertEqualObjects(video.identifier, @"9TTioMbNT9I");
 		XCTAssertEqualObjects(video.title, @"Super Mario Bros Theme Song on Wine Glasses and a Frying Pan (슈퍼 마리오 브라더스 - スーパーマリオブラザーズ - 超級瑪莉)");
@@ -34,15 +36,16 @@
 		XCTAssertNotNil(video.largeThumbnailURL);
 		XCTAssertTrue(video.streamURLs.count > 0);
 		XCTAssertTrue(video.duration > 0);
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testMobileRestrictedVideo
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"JHaA9bKi-xs" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"JHaA9bKi-xs" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
 		XCTAssertNotNil(video.smallThumbnailURL);
@@ -53,15 +56,16 @@
 		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop) {
 			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
 		}];
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
-- (void) testLiveVideo
+- (void) testLiveVideo_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"xrM34fdmloc" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"xrM34fdmloc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
 		XCTAssertNotNil(video.smallThumbnailURL);
@@ -70,15 +74,16 @@
 		XCTAssertEqual(video.streamURLs.count, 1U);
 		XCTAssertTrue(video.duration > 0);
 		XCTAssertNotNil(video.streamURLs[XCDYouTubeVideoQualityHTTPLiveStreaming]);
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 - (void) testDVRVideo
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"H7iQ4sAf0OE" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"H7iQ4sAf0OE" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
 		XCTAssertNotNil(video.smallThumbnailURL);
@@ -87,106 +92,129 @@
 		XCTAssertEqual(video.streamURLs.count, 1U);
 		XCTAssertTrue(video.duration > 0);
 		XCTAssertNotNil(video.streamURLs[XCDYouTubeVideoQualityHTTPLiveStreaming]);
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testRestrictedVideo
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"1kIsylLeHHU" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"1kIsylLeHHU" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
-		XCTAssertEqualObjects(error.localizedDescription, @"This video contains content from Youtube test content owner, who has blocked it on copyright grounds.");
-		[monitor signal];
+		XCTAssertEqualObjects(error.localizedDescription, @"This video is currently unavailable. We are working to bring it back.");
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testRemovedVideo
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"BXnA9FjvLSU" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"BXnA9FjvLSU" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRemovedVideo);
 		XCTAssertEqualObjects(error.localizedDescription, @"\"9/11 The F...\" This video is no longer available due to a copyright claim by Digital Rights Group Ltd.");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testGeoblockedVideo
+{
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"vwkFTztnl7Y" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNil(video);
+		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
+		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
+		XCTAssertEqualObjects(error.localizedDescription, @"The uploader has not made this video available in your country.");
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testInvalidVideoIdentifier
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"tooShort" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"tooShort" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorInvalidVideoIdentifier);
 		XCTAssertEqualObjects(error.localizedDescription, @"Invalid parameters.");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testNonExistentVideoIdentifier
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"xxxxxxxxxxx" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"xxxxxxxxxxx" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRemovedVideo);
 		XCTAssertEqualObjects(error.localizedDescription, @"This video does not exist.");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testFrenchClient
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[[XCDYouTubeClient alloc] initWithLanguageIdentifier:@"fr"] getVideoWithIdentifier:@"xxxxxxxxxxx" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[[XCDYouTubeClient alloc] initWithLanguageIdentifier:@"fr"] getVideoWithIdentifier:@"xxxxxxxxxxx" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRemovedVideo);
 		XCTAssertEqualObjects(error.localizedDescription, @"Cette vidéo n'existe pas.");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testNilVideoIdentifier
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:nil completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:nil completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorInvalidVideoIdentifier);
 		XCTAssertEqualObjects(error.localizedDescription, @"Invalid parameters.");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testSpaceVideoIdentifier
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@" " completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@" " completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorInvalidVideoIdentifier);
 		XCTAssertEqualObjects(error.localizedDescription, @"Invalid parameters.");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
-- (void) testConnectionError
+// Disable internet connection before running
+- (void) testConnectionError_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"EdeVaT-zZt4" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"EdeVaT-zZt4" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorNetwork);
@@ -194,32 +222,35 @@
 		NSError *underlyingError = error.userInfo[NSUnderlyingErrorKey];
 		XCTAssertEqualObjects(underlyingError.domain, NSURLErrorDomain);
 		XCTAssertEqual(underlyingError.code, NSURLErrorNotConnectedToInternet);
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 - (void) testUsingClientOnNonMainThread
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
 	dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
 		XCTAssertFalse([NSThread isMainThread]);
-		[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"EdeVaT-zZt4" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+		[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"EdeVaT-zZt4" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+		{
 			XCTAssertTrue([NSThread isMainThread]);
-			[monitor signal];
+			[expectation fulfill];
 		}];
 	});
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testCancelingOperation
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	id<XCDYouTubeOperation> operation = [[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"EdeVaT-zZt4" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	id<XCDYouTubeOperation> operation = [[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"EdeVaT-zZt4" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTFail();
 	}];
+	[expectation performSelector:@selector(fulfill) withObject:nil afterDelay:0.2];
 	[operation cancel];
-	XCTAssertFalse([monitor waitWithTimeout:0.2]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 - (void) testNilCompletionHandler