+++ XCDYouTubeKit Tests/XCDYouTubeKitTestCase.m
@@ -8,15 +8,51 @@
 - (void) setUpTestWithSelector:(SEL)selector;
 @end
 
+@interface NSURLRequest (Private)
++ (void) setAllowsAnyHTTPSCertificate:(BOOL)allowsAnyHTTPSCertificate forHost:(NSString *)host;
+@end
+
 @interface XCDYouTubeKitTestCase ()
 @property NSURL *cassetteURL;
 @end
 
+static NSString *const offlineSuffix = @"_offline";
+
 @implementation XCDYouTubeKitTestCase
+
++ (void) setUp
+{
+	NSDictionary *environment = NSProcessInfo.processInfo.environment;
+	if ([environment[@"VCR_CASSETTES_DIRECTORY"] boolValue] || [environment[@"ONLINE_TESTS"] boolValue])
+	{
+		[NSURLRequest setAllowsAnyHTTPSCertificate:YES forHost:@"www.youtube.com"];
+		[NSURLRequest setAllowsAnyHTTPSCertificate:YES forHost:@"s.ytimg.com"];
+	}
+}
+
+// When running tests with the `ONLINE_TESTS` environment variable, tests whose
+// selector ends with `_offline` are not executed.
+// Also, VCR is not used for replaying when `ONLINE_TESTS` is true.
++ (NSArray *) testInvocations
+{
+	BOOL onlineTests = [[[[NSProcessInfo processInfo] environment] objectForKey:@"ONLINE_TESTS"] boolValue];
+	if (!onlineTests)
+		return [super testInvocations];
+	
+	NSMutableArray *testInvocations = [NSMutableArray new];
+	for (NSInvocation *invocation in [super testInvocations])
+	{
+		if (![NSStringFromSelector(invocation.selector) hasSuffix:offlineSuffix])
+			[testInvocations addObject:invocation];
+	}
+	return [testInvocations copy];
+}
 
 - (void) setUpTestWithSelector:(SEL)selector
 {
 	[super setUpTestWithSelector:selector];
+	
+	BOOL onlineTests = [[[[NSProcessInfo processInfo] environment] objectForKey:@"ONLINE_TESTS"] boolValue];
 	
 	NSString *cassettesDirectory = [[[NSProcessInfo processInfo] environment] objectForKey:@"VCR_CASSETTES_DIRECTORY"];
 	cassettesDirectory = [cassettesDirectory stringByAppendingPathComponent:NSStringFromClass(self.class)];
@@ -28,7 +64,14 @@
 	}
 	else
 	{
-		self.cassetteURL = [[NSBundle bundleForClass:self.class] URLForResource:NSStringFromSelector(selector) withExtension:@"json" subdirectory:[@"Cassettes" stringByAppendingPathComponent:NSStringFromClass(self.class)]];
+		if (onlineTests)
+			return;
+		
+		NSString *testName = NSStringFromSelector(selector);
+		if ([testName hasSuffix:offlineSuffix])
+			testName = [testName substringToIndex:testName.length - offlineSuffix.length];
+		
+		self.cassetteURL = [[NSBundle bundleForClass:self.class] URLForResource:testName withExtension:@"json" subdirectory:[@"Cassettes" stringByAppendingPathComponent:NSStringFromClass(self.class)]];
 		XCTAssertNotNil(self.cassetteURL);
 		[VCR loadCassetteWithContentsOfURL:self.cassetteURL];
 		[VCR setReplaying:YES];