+++ XCDYouTubeKit Tests/XCDYouTubeProtectedVideosTestCase.m
@@ -6,35 +6,16 @@
 
 #import <XCDYouTubeKit/XCDYouTubeClient.h>
 
-@interface NSURLRequest (Private)
-+ (void) setAllowsAnyHTTPSCertificate:(BOOL)allowsAnyHTTPSCertificate forHost:(NSString *)host;
-@end
-
 @interface XCDYouTubeProtectedVideosTestCase : XCDYouTubeKitTestCase
 @end
 
 @implementation XCDYouTubeProtectedVideosTestCase
 
-- (void) setUp
+- (void) testAgeRestrictedVideo
 {
-	[super setUp];
-	
-	if (NSProcessInfo.processInfo.environment[@"VCR_CASSETTES_DIRECTORY"])
-		[NSURLRequest setAllowsAnyHTTPSCertificate:YES forHost:@"www.youtube.com"];
-}
-
-- (void) tearDown
-{
-	[super tearDown];
-	
-	if (NSProcessInfo.processInfo.environment[@"VCR_CASSETTES_DIRECTORY"])
-		[NSURLRequest setAllowsAnyHTTPSCertificate:NO forHost:@"www.youtube.com"];
-}
-
-- (void) testProtectedVEVOVideo
-{
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"rId6PKlDXeU" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"zKovmts2KSk" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
 		XCTAssertNotNil(video.smallThumbnailURL);
@@ -42,37 +23,20 @@
 		XCTAssertNotNil(video.largeThumbnailURL);
 		XCTAssertTrue(video.streamURLs.count > 0);
 		XCTAssertTrue(video.duration > 0);
-		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop) {
+		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop)
+		{
 			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
 		}];
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
-- (void) testProtectedVideoWithDollarSignature
+- (void) testProtectedVEVOVideo1
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
-		XCTAssertNil(error);
-		XCTAssertNotNil(video.title);
-		XCTAssertNotNil(video.smallThumbnailURL);
-		XCTAssertNotNil(video.mediumThumbnailURL);
-		XCTAssertNil(video.largeThumbnailURL);
-		XCTAssertTrue(video.streamURLs.count > 0);
-		XCTAssertTrue(video.duration > 0);
-		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop) {
-			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
-		}];
-		[monitor signal];
-	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
-}
-
-- (void) testProtectedVideoWithJavaScriptFunctionsInVarScope
-{
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Ntn1-SocNiY" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"rId6PKlDXeU" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
 		XCTAssertNotNil(video.smallThumbnailURL);
@@ -80,84 +44,153 @@
 		XCTAssertNotNil(video.largeThumbnailURL);
 		XCTAssertTrue(video.streamURLs.count > 0);
 		XCTAssertTrue(video.duration > 0);
-		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop) {
+		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop)
+		{
 			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
 		}];
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testProtectedVEVOVideo2
+{
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNil(error);
+		XCTAssertNotNil(video.title);
+		XCTAssertNotNil(video.smallThumbnailURL);
+		XCTAssertNotNil(video.mediumThumbnailURL);
+		XCTAssertNotNil(video.largeThumbnailURL);
+		XCTAssertTrue(video.streamURLs.count > 0);
+		XCTAssertTrue(video.duration > 0);
+		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop)
+		{
+			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
+		}];
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testProtectedVEVOVideo3
+{
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Ntn1-SocNiY" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNil(error);
+		XCTAssertNotNil(video.title);
+		XCTAssertNotNil(video.smallThumbnailURL);
+		XCTAssertNotNil(video.mediumThumbnailURL);
+		XCTAssertNotNil(video.largeThumbnailURL);
+		XCTAssertTrue(video.streamURLs.count > 0);
+		XCTAssertTrue(video.duration > 0);
+		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop)
+		{
+			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
+		}];
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testAgeRestrictedVEVOVideo
+{
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"07FYdnEawAQ" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNil(error);
+		XCTAssertNotNil(video.title);
+		XCTAssertNotNil(video.smallThumbnailURL);
+		XCTAssertNotNil(video.mediumThumbnailURL);
+		XCTAssertNotNil(video.largeThumbnailURL);
+		XCTAssertTrue(video.streamURLs.count > 0);
+		XCTAssertTrue(video.duration > 0);
+		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop)
+		{
+			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
+		}];
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 // With Charles
 //   * Enable SSL proxying for *.youtube.com
 //   * Tools -> Black List... -> Add host:www.youtube.com and path:watch to simulate connection error on the web page
-- (void) testProtectedVideoWithWebPageConnectionError
+- (void) testProtectedVideoWithWebPageConnectionError_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
 		XCTAssertEqualObjects(error.localizedDescription, @"This video contains content from WMG. It is restricted from playback on certain sites. Watch on YouTube");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 // With Charles: Tools -> Black List... -> Add `s.ytimg.com` to simulate connection error on the player script
-- (void) testProtectedVideoWithPlayerScriptConnectionError
+- (void) testProtectedVideoWithPlayerScriptConnectionError_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
 		XCTAssertEqualObjects(error.localizedDescription, @"This video contains content from WMG. It is restricted from playback on certain sites. Watch on YouTube");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 // Edit testProtectedVideoWithoutSignatureFunction.json by replacing `signature=` with `signaturX=`
-- (void) testProtectedVideoWithoutSignatureFunction
+- (void) testProtectedVideoWithoutSignatureFunction_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
 		XCTAssertEqualObjects(error.localizedDescription, @"This video contains content from WMG. It is restricted from playback on certain sites. Watch on YouTube");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 // Edit testProtectedVideoWithBrokenSignatureFunction.json by returning null in the signature function
-- (void) testProtectedVideoWithBrokenSignatureFunction
+- (void) testProtectedVideoWithBrokenSignatureFunction_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
 		XCTAssertEqualObjects(error.localizedDescription, @"This video contains content from WMG. It is restricted from playback on certain sites. Watch on YouTube");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 // Edit testProtectedVideoWithoutJavaScriptPlayerURL.json by replacing `\"js\":` with `\"xs\":`
-- (void) testProtectedVideoWithoutJavaScriptPlayerURL
+- (void) testProtectedVideoWithoutJavaScriptPlayerURL_offline
 {
-	TRVSMonitor *monitor = [TRVSMonitor monitor];
-	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error) {
+	XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"Pgum6OT_VH8" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorRestrictedPlayback);
 		XCTAssertEqualObjects(error.localizedDescription, @"This video contains content from WMG. It is restricted from playback on certain sites. Watch on YouTube");
-		[monitor signal];
+		[expectation fulfill];
 	}];
-	XCTAssertTrue([monitor waitWithTimeout:10]);
+	[self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 @end