+++ XCDYouTubeKit/XCDYouTubeError.h
@@ -14,7 +14,7 @@ extern NSString *const XCDYouTubeVideoErrorDomain;
  */
 typedef NS_ENUM(NSInteger, XCDYouTubeErrorCode) {
 	/**
-	 *  Returned if no suitable video stream is available.
+	 *  Returned when no suitable video stream is available.
 	 */
 	XCDYouTubeErrorNoStreamAvailable      = -2,
 	
@@ -29,12 +29,12 @@ typedef NS_ENUM(NSInteger, XCDYouTubeErrorCode) {
 	XCDYouTubeErrorInvalidVideoIdentifier = 2,
 	
 	/**
-	 *  Returned when the video has been removed as a violation of YouTube's policy or if the video does not exist.
+	 *  Returned when the video has been removed as a violation of YouTube's policy or when the video does not exist.
 	 */
 	XCDYouTubeErrorRemovedVideo           = 100,
 	
 	/**
-	 *  Returned when the video is not playable because of legal reasons or the this is a private video.
+	 *  Returned when the video is not playable because of legal reasons or when the video is private.
 	 */
 	XCDYouTubeErrorRestrictedPlayback     = 150
 };