+++ XCDYouTubeKit/XCDYouTubeVideoWebpage.m
@@ -14,6 +14,7 @@
 	NSDictionary *_playerConfiguration;
 	NSDictionary *_videoInfo;
 	NSURL *_javaScriptPlayerURL;
+	BOOL _isAgeRestricted;
 }
 
 - (instancetype) initWithData:(NSData *)data response:(NSURLResponse *)response
@@ -37,7 +38,8 @@
 		CFStringEncoding encoding = CFStringConvertIANACharSetNameToEncoding((__bridge CFStringRef)self.response.textEncodingName ?: CFSTR(""));
 		NSString *html = CFBridgingRelease(CFStringCreateWithBytes(kCFAllocatorDefault, [self.data bytes], (CFIndex)[self.data length], encoding != kCFStringEncodingInvalidId ? encoding : kCFStringEncodingISOLatin1, false));
 		NSRegularExpression *playerConfigRegularExpression = [NSRegularExpression regularExpressionWithPattern:@"ytplayer.config\\s*=\\s*(\\{.*?\\});" options:NSRegularExpressionCaseInsensitive error:NULL];
-		[playerConfigRegularExpression enumerateMatchesInString:html options:(NSMatchingOptions)0 range:NSMakeRange(0, html.length) usingBlock:^(NSTextCheckingResult *result, NSMatchingFlags flags, BOOL *stop) {
+		[playerConfigRegularExpression enumerateMatchesInString:html options:(NSMatchingOptions)0 range:NSMakeRange(0, html.length) usingBlock:^(NSTextCheckingResult *result, NSMatchingFlags flags, BOOL *stop)
+		{
 			NSString *configString = [html substringWithRange:[result rangeAtIndex:1]];
 			NSDictionary *playerConfiguration = [NSJSONSerialization JSONObjectWithData:[configString dataUsingEncoding:NSUTF8StringEncoding] options:(NSJSONReadingOptions)0 error:NULL];
 			if ([playerConfiguration isKindOfClass:[NSDictionary class]])
@@ -59,7 +61,8 @@
 		if ([args isKindOfClass:[NSDictionary class]])
 		{
 			NSMutableDictionary *info = [NSMutableDictionary new];
-			[args enumerateKeysAndObjectsUsingBlock:^(id key, id value, BOOL *stop) {
+			[args enumerateKeysAndObjectsUsingBlock:^(id key, id value, BOOL *stop)
+			{
 				if ([value isKindOfClass:[NSString class]] || [value isKindOfClass:[NSNumber class]])
 					info[key] = [value description];
 			}];
@@ -84,6 +87,18 @@
 		}
 	}
 	return _javaScriptPlayerURL;
+}
+
+- (BOOL) isAgeRestricted
+{
+	if (!_isAgeRestricted)
+	{
+		NSData *openGraphAgeRestriction = [@"og:restrictions:age" dataUsingEncoding:NSUTF8StringEncoding];
+		NSDataSearchOptions options = (NSDataSearchOptions)0;
+		NSRange range = NSMakeRange(0, self.data.length);
+		_isAgeRestricted = [self.data rangeOfData:openGraphAgeRestriction options:options range:range].location != NSNotFound;
+	}
+	return _isAgeRestricted;
 }
 
 @end